backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
cryptography==41.0.7
watchdog==3.0.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
sqlalchemy==2.0.23
aiosqlite==0.19.0
pydantic==2.5.0
pydantic-settings==2.1.0
prometheus-client==0.19.0
